refactor: change Item component to Card (#49)
